% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.R
\name{procrustes_distance}
\alias{procrustes_distance}
\title{Procrustes distance between two subspaces}
\usage{
procrustes_distance(U, V)
}
\arguments{
\item{U, V}{Basis matrices for the two subspaces.}
}
\value{
Nonnegative scalar distance.
}
\description{
Computes the orthogonal Procrustes distance between subspaces spanned by \code{U}
and \code{V} using singular values of \verb{U'V}.
}
\examples{
\dontrun{
d <- procrustes_distance(U, V)
}

}
\seealso{
\code{\link[=subspace_principal_angles]{subspace_principal_angles()}} for computing angles between subspaces

Other subspace analysis: 
\code{\link{subspace_principal_angles}()},
\code{\link{whitened_matrix}()},
\code{\link{whitened_svd}()}
}
\concept{subspace analysis}
